requirements.txt
PySide6>=6.6.0
anthropic>=0.18.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
qasync>=0.27.0
